requirements.txt
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy
Python-Dotenv
PY-BCRYPT